Move suite execution & status results to api(#50)
