+++ api/communication-api/src/main/java/com/cognifide/aet/communication/api/suiteexecution/ProcessingStatus.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.aet.executor;
+package com.cognifide.aet.communication.api.suiteexecution;
 
 public enum ProcessingStatus {
