+++ test-executor/src/main/java/com/cognifide/aet/executor/SuiteExecutor.java
@@ -27,6 +27,9 @@ import com.google.common.collect.Sets;
 import com.cognifide.aet.communication.api.metadata.Suite;
 import com.cognifide.aet.communication.api.metadata.ValidatorException;
 import com.cognifide.aet.communication.api.queues.JmsConnection;
+import com.cognifide.aet.communication.api.suiteexecution.ProcessingStatus;
+import com.cognifide.aet.communication.api.suiteexecution.SuiteExecutionResult;
+import com.cognifide.aet.communication.api.suiteexecution.SuiteStatusResult;
 import com.cognifide.aet.executor.common.SuiteFactory;
 import com.cognifide.aet.executor.model.CollectorStep;
 import com.cognifide.aet.executor.model.ComparatorStep;
