+++ test-executor/src/main/java/com/cognifide/aet/executor/SuiteServlet.java
@@ -19,6 +19,8 @@ package com.cognifide.aet.executor;
 
 import com.google.gson.Gson;
 
+import com.cognifide.aet.communication.api.suiteexecution.SuiteExecutionResult;
+
 import org.apache.commons.fileupload.FileItemIterator;
 import org.apache.commons.fileupload.FileItemStream;
 import org.apache.commons.fileupload.FileUploadException;
